refactor: move AutocompleteInput to regular components
