+++ components/advanced-search/MultiSearchInput.vue
@@ -84,7 +84,7 @@
 
 <script lang="ts">
 import Vue, { PropType } from "vue";
-import AutocompleteInput from "@/components/advanced-search/AutocompleteInput.vue";
+import AutocompleteInput from "@/components/AutocompleteInput.vue";
 import StyledSelect from "@/components/StyledSelect.vue";
 
 type MultiSearchInputValue = { value: string; operator: string }[];
